refactor(agent-core): move the entry-inputSchema warning into check() so MCP + canvas inherit it (SAP-2227)
Addresses the blocking review on #469: with the warning layered in the CLI's runCheck(),
the sapiom_dev_agents_check MCP tool and the dashboard canvas — both of which call
@sapiom/agent-core check() directly — never saw it. That MCP tool is exactly the surface
the primer points coding agents at, so the ticket's "human AND coding-agent" goal was unmet.

- generate the warning inside check(), pushed to the shared `warnings` array, so the CLI,
  the MCP tool, and the harness canvas all inherit it (CLI reverts to a plain relay)
- move the unit test back to packages/agent-core (check.spec.ts)
- the harness canvas now surfaces the note on schema-less agents; update the order-triage
  fixture expectation (that fixture ships no inputSchema by design — a monorepo zod hazard)

Rebased onto main, which now carries SAP-2226 (#471): its polished default-template entry
schema supersedes the one this branch added, so that template edit is dropped.
